requirements.txt
python-telegram-bot==21.5
groq==0.11.0
google-generativeai==0.8.0
python-docx==1.1.2
python-dotenv==1.0.1
httpx==0.27.2
